youtube: fix playback issue
To avoid showing undesired video suggestions whenever the player pauses
or ends, we have configured the YouTube player to destroy in case of
such events.

The issue was that if the user was selecting a video to play while
another was already playing, the player would destroy (as the
ENDED event would be triggered for the first video) and the next video
would never play.

The quick and dirty way to fix this issue is to instantiate a new player
for each videoId. It could be cleaner to have some proper state
management in the web app, but this is not my primary focus right now.